+++ tests/testthat/test_bug_154.R
@@ -5,15 +5,13 @@ test_that("Bug 154 is fixed: rule columns populate correctly", {
   
   results_1 <- autospc(bug_148_data %>%
                                dplyr::filter(x <= 287L),
-                             chart_type = "XMR",
-                             plot_chart = FALSE,
-                             show_mr = FALSE)
+                             chart_type = "X",
+                             plot_chart = FALSE)
   
   results_2 <- autospc(bug_148_data %>%
                                dplyr::filter(x <= 288L),
-                             chart_type = "XMR",
-                             plot_chart = FALSE,
-                             show_mr = FALSE)
+                             chart_type = "X",
+                             plot_chart = FALSE)
   
   rule2_NAs_1 <- results_1 %>% 
     dplyr::summarise(num_nas = sum(is.na(rule2))) %>%