+++ tests/testthat/test_x_break.R
@@ -2,18 +2,15 @@
 test_that("Integer x breaks are as specified by x_break", {
   
   plot_10 <- autospc(example_series_1,
-                     chart_type = "XMR",
-                     show_mr = FALSE,
+                     chart_type = "X",
                      x_break = 10)
   
   plot_20 <- autospc(example_series_1,
-                     chart_type = "XMR",
-                     show_mr = FALSE,
+                     chart_type = "X",
                      x_break = 20)
   
   plot_5<- autospc(example_series_1,
-                   chart_type = "XMR",
-                   show_mr = FALSE,
+                   chart_type = "X",
                    x_break = 5)
   
   n_x_breaks_10 <- length(as.numeric(na.omit(